Last fixes for examples to compile
